% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_sats.R
\name{align.sats}
\alias{align.sats}
\title{Align a Raster Time Series Stemming from Two Different Satellites}
\usage{
align.sats(x, sat.order, to.from, mean.only = FALSE, filename = "", ...)
}
\arguments{
\item{x}{Raster time series. Either a SpatRaster, a RasterStack, or a list of SpatRasters, RasterLayers, or file paths. The two time series must have same projection, extent, and resolution.}

\item{sat.order}{A vector denoting the satellite each raster stems from.}

\item{to.from}{Vector declaring which time series will be aligned with the other. The "from" series will be aligned with the "to" series. The elements of \code{to.from} must match the unique values of \code{sat.order}.}

\item{mean.only}{If \code{TRUE}, only the mean of the "from" series will change. If \code{FALSE}, both the mean and the standard deviation will be adjusted.}

\item{filename}{Output file path.}

\item{...}{Additional arguments to pass to \code{writeRaster}.}
}
\value{
A SpatRaster of the aligned time series.
}
\description{
Ensures that, on average, there will be no difference in mean or standard deviation between two overlapping raster time series obtained by different satellites.
}
\examples{
#load data
data(rts_ex)
data(so_ex)

#before
avgs <- sapply(1:dim(rts_ex)[3], function(i) mean(rts_ex[,,i], na.rm = TRUE))
ggplot2::ggplot() +
ggplot2::geom_point(ggplot2::aes(x = 1:length(avgs), y = avgs, col = factor(so_ex)))

#after
rts_ex <- align.sats(rast(rts_ex), so_ex, c("HLSL","HLSS"))
avgs <- sapply(1:dim(rts_ex)[3], function(i) mean(rts_ex[,,i], na.rm = TRUE))
ggplot2::ggplot() +
ggplot2::geom_point(ggplot2::aes(x = 1:length(avgs), y = avgs, col = factor(so_ex)))
}
