% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arr_ex}
\alias{arr_ex}
\title{Example 3-Dimensional Array}
\format{
An array with dim = c(5,5,300).
}
\usage{
arr_ex
}
\description{
Example array constructed from \code{ts_ex}. Is converted to SpatRaster and used as input to \code{detect.changes}.
}
\keyword{datasets}
