% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scale.R
\name{check.scale}
\alias{check.scale}
\title{Check and Correct Raster Scale}
\usage{
check.scale(raster, scalar, tolerance = 1, filename = "", ...)
}
\arguments{
\item{raster}{A SpatRaster, RasterLayer, or file path.}

\item{scalar}{Target scale. If the average value of the raster differs from this number too greatly, the raster will be multiplied by this number.}

\item{tolerance}{Tolerated difference in significant figures between scalar and average raster value. Exceeding this value results in rescaling.}

\item{filename}{Output file path.}

\item{...}{Additional argument to pass to \code{writeRaster}.}
}
\value{
A SpatRaster.
}
\description{
Compare the difference in significant digits between the average value of a raster and a scalar. If the difference in significant digits exceeds a threshold, the raster is multiplied by the scalar.
}
\examples{
set.seed(1)
m <- matrix(rnorm(100), nrow = 10)
r <- rast(m)
mean(values(r))

#check r for proper scale, rescale if needed
r <- check.scale(r, 10000, 1)
mean(values(r))
}
