% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cl_ex}
\alias{cl_ex}
\title{Example Change List}
\format{
List of lists. \verb{$info} contains raster properties, while \verb{$changes} details changes detected by \code{detect.changes}.
}
\usage{
cl_ex
}
\description{
Output of \code{detect.changes}; used as input for \code{measure.change}.
}
\keyword{datasets}
