% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_change.R
\name{detect.change}
\alias{detect.change}
\title{Detect Changes in a Raster Time Series}
\usage{
detect.change(x, t.Gregorian, t.fit = 365, min.obs.fit = 1, filename = "", ...)
}
\arguments{
\item{x}{Raster time series. Either a SpatRaster, a RasterStack, or a list of SpatRasters, RasterLayers, or file paths.}

\item{t.Gregorian}{Vector of dates associated with the time series. Use \%Y-\%m-\%d format.}

\item{t.fit}{Time window used to fit new models, expressed in number of (whole) days.}

\item{min.obs.fit}{Minimum number of observations to fit an initial model. If fewer than \code{min.obs.fit} observations in the initial fitting period are not \code{NA}, the pixel will be skipped. Does not apply to subsequent models.}

\item{filename}{Output file path. Valid extensions are .rdata, .rds, .json, and .yaml.}

\item{...}{Additional arguments to pass to \code{detect.change.harmonic}.}
}
\value{
A list containing two sublists. \verb{$info} contains the raster's dimensions, coordinate reference system, extent, and resolution, while \verb{$changes} pairs the output of \code{detect.change.harmonic} with the index of the corresponding pixel. Only pixels for which a change occurred are included in \verb{$changes}, and if no changes occurred or all pixels were skipped, \verb{$changes} will simply be \code{NA}.
}
\description{
Applies a change detection algorithm to all pixels in a raster time series.
}
\examples{
#load/create data
data(arr_ex)
data(ts_ex)

#detect changes
cl_ex <- detect.change(rast(arr_ex), tg_ex, flag.stat = "iqr")
print("info")
cl_ex$info
print("changes")
cl_ex$changes[[1]]
print(paste0("length: ", length(cl_ex$changes)))
}
