% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fm_ex}
\alias{fm_ex}
\title{Example Fmask raster}
\format{
A RasterLayer with 3660 rows and 3660 columns.
\itemize{
\item{satellite:}{Landsat 8/9}
\item{date:}{2014.08.25}
\item{time:}{18:05:15}
\item{crs:}{+proj=utm +zone=12 +ellps=WGS84 +units=m +no_defs}
\item{res:}{30}
\item{ext:}{443880, 454860, 3923160, 3934140 (xmin, xmax, ymin, ymax)}
}
}
\source{
AppEEARS Team. (2022). Application for Extracting and Exploring Analysis Ready Samples (AppEEARS). Ver. 3.19. NASA EOSDIS Land Processes Distributed Active Archive Center (LP DAAC), USGS/Earth Resources Observation and Science (EROS) Center, Sioux Falls, South Dakota, USA. Accessed December 9, 2022. \href{https://appeears.earthdatacloud.nasa.gov}{https://appeears.earthdatacloud.nasa.gov}.

Earthdata Search. 2019. Greenbelt, MD: Earth Science Data and Information System (ESDIS) Project, Earth Science Projects Division (ESPD), Flight Projects Directorate, Goddard Space Flight Center (GSFC) National Aeronautics and Space Administration (NASA). URL:  \href{https://search.earthdata.nasa.gov}{https://search.earthdata.nasa.gov}.

Masek, J., Ju, J., Roger, J., Skakun, S., Vermote, E., Claverie, M., Dungan, J., Yin, Z., Freitag, B., Justice, C. (2021). HLS Operational Land Imager Surface Reflectance and TOA Brightness Daily Global 30m v2.0. NASA EOSDIS Land Processes DAAC. Accessed 2022-12-09 from \href{https://doi.org/10.5067/HLS/HLSL30.002}{https://doi.org/10.5067/HLS/HLSL30.002}. Accessed December 9, 2022.
}
\usage{
fm_ex
}
\description{
Example Fmask raster
}
\keyword{datasets}
