% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qa.R
\name{make.qa}
\alias{make.qa}
\title{Create a QA Layer for Harmonized Landsat Sentinel-2 Data}
\source{
The Fmask lookup table used in this function can be obtained by downloading HLS data through \href{https://appeears.earthdatacloud.nasa.gov/}{AppEEARS}.
}
\usage{
make.qa(x, keep = NULL, filename = "", ...)
}
\arguments{
\item{x}{A HLS Fmask layer. Either a SpatRaster, RasterLayer, or file path.}

\item{keep}{Character vector of pixel classifications to be excluded from masking. Options are "Water," "Snow/ice," "Cloud shadow," "Adjacent to cloud/shadow," and "Cloud".}

\item{filename}{Output file path.}

\item{...}{Additional arguments to pass to \code{writeRaster}.}
}
\value{
A SpatRaster. Non-NA pixel values are 0 (Low aerosol), 1 (Moderate aerosol), or 2 (High aerosol).
}
\description{
Creates a QA layer for a Harmonized Landsat Sentinel-2 (HLSL/HLSS) image. Masks clouds, snow, etc. while ranking the unmasked pixels according to aerosol level.
}
\examples{
data(fm_ex)
plot(fm_ex)

#QA layer with all categories masked
qa <- make.qa(fm_ex)
plot(qa)

#QA layer with cloud shadow and adjacent pixels retained
qa <- make.qa(fm_ex, keep = c("Cloud shadow", "Adjacent to cloud/shadow"))
plot(qa)
}
