% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_clean.R
\name{mask.clean}
\alias{mask.clean}
\title{Mask, Remove Outliers from, and Rescale a Raster}
\usage{
mask.clean(
  raster,
  layer = NULL,
  layer.type = NULL,
  scalar = NULL,
  tolerance = 1,
  filename = "",
  ...
)
}
\arguments{
\item{raster}{A SpatRaster, RasterLayer, or file path.}

\item{layer}{Layer for masking. Either a QA generated by \code{make.qa}, or an Fmask layer, in which case a QA layer will be generated and used for masking, but won't be included in the output.}

\item{layer.type}{Tells the function whether \code{layer} is a QA or Fmask layer. Valid options are "QA" or "FM."}

\item{scalar}{Target scale. If the average value of the raster differs from this number too greatly, the raster will be multiplied by this number.}

\item{tolerance}{Tolerated difference in significant figures between scalar and average raster value. Exceeding this value results in rescaling.}

\item{filename}{Output file path.}

\item{...}{Additional argument to pass to \code{writeRaster} and \code{check.scale}.}
}
\value{
A SpatRaster.
}
\description{
Combined function for pre-processing tasks. If only the first argument is supplied, \code{mask.clean} is identical to \code{remove.outliers}. However, additional parameters allow the user to also mask the input raster with a QA layer (which can be generated if an HLS Fmask layer is supplied), as well as rescale it if it is not properly scaled.
}
\examples{
#Load/create data
set.seed(1)
data(qa_ex)
r <- rast(matrix(rnorm(100), nrow = 10))

#Compare before and after `mask.clean`
mean(values(r))
plot(r, main = "Raster")
plot(qa_ex, main = "QA layer")
r <- mask.clean(r, qa_ex, "QA", 10000)
mean(values(r), na.rm = TRUE)
plot(r, main = "Masked & Cleaned Raster")
}
