% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_change.R
\name{measure.change}
\alias{measure.change}
\title{Measure Detected Changes}
\usage{
measure.change(x, t.Gregorian, t.sub = NULL, filename = "")
}
\arguments{
\item{x}{A list created by \code{detect.change}.}

\item{t.Gregorian}{Vector of dates associated with the time series. Use \%Y-\%m-\%d format.}

\item{t.sub}{Optional vector used to subset \code{t.Gregorian}. Should contain only a start and end index, i.e. \code{c(1,10)}.}

\item{filename}{Output file path. Valid extensions are .rdata, .rds, .json, and .yaml.}
}
\value{
Modifies the input list. \verb{$dates} will be added to \code{x$info}; \verb{$diff.raw} and \verb{$diff.pct} will be added to \code{x$changes}. List elements corresponding to initial models are removed. If \code{t.sub} is not \code{NULL}, changes occurring outside of the subset will also be removed.
}
\description{
Calculates the raw and percent difference between model coefficients for changes detected by \code{detect.change} and stores them in a list.
}
\examples{
#load data
data(cl_ex)
data(tg_ex)

#measure change for whole period
mc_ex <- measure.change(cl_ex, tg_ex)
mc_ex[[1]]$dates
head(rlist::list.stack(mc_ex[[2]]))

#measure change for slice
mc_ex <- measure.change(cl_ex, tg_ex, c(50, 150))
mc_ex[[1]]$dates
head(rlist::list.stack(mc_ex[[2]]))
}
