% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qa_ex}
\alias{qa_ex}
\title{Example QA Layer}
\format{
A RasterLayer with 10 rows and 10 columns.
}
\usage{
qa_ex
}
\description{
Example QA Layer generated with \code{make.qa}. Used as input for \code{mask.clean}.
}
\keyword{datasets}
