% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove.outliers}
\alias{remove.outliers}
\title{Remove Extreme Values from a Raster}
\usage{
remove.outliers(raster, filename = "", ...)
}
\arguments{
\item{raster}{A SpatRaster, RasterLayer, or file path.}

\item{filename}{Output file path.}

\item{...}{Additional argument to pass to \code{writeRaster}.}
}
\value{
A SpatRaster.
}
\description{
Remove Extreme Values from a Raster
}
\examples{
set.seed(1)
m <- matrix(rbinom(100, 10, 0.5), nrow = 10)
s <- sample(1:100, 5)
m[s] <- 100

r <- rast(m)
plot(r)

r <- remove.outliers(r)
plot(r)
}
