% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rts_ex}
\alias{rts_ex}
\title{Example Raster Time Series}
\format{
A RasterLayer with dim = c(5,5,300).
}
\usage{
rts_ex
}
\description{
Same as \code{arr_ex}, but every other raster has a higher mean and variance. Meant to simulate differences in time series from two satellites.
}
\keyword{datasets}
