% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{so_ex}
\alias{so_ex}
\title{Example Satellite Order}
\format{
Character vector of length 300.
}
\usage{
so_ex
}
\description{
Used by \code{align.sats} to determine which rasters are from which satellites.
}
\keyword{datasets}
