% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tg_ex}
\alias{tg_ex}
\title{Example Gregorian Dates}
\format{
Vector of 300 dates spanning Jan 18, 2020 to Apr 28, 2021.
}
\usage{
tg_ex
}
\description{
Example Gregorian Dates
}
\keyword{datasets}
